+++ Sasquatch/Sasquatch/ViewControllers/MSMainViewController.swift
@@ -7,6 +7,8 @@ import UIKit
 let kMSDefaultDatabaseSize = 10 * 1024 * 1024
 let acProdLogUrl = "https://in.appcenter.ms"
 let ocProdLogUrl = "https://mobile.events.data.microsoft.com"
+let kMSARefreshTokenKey = "MSARefreshToken"
+let kMSATokenKey = "MSAToken"
 
 class MSMainViewController: UITableViewController, AppCenterProtocol {
   
@@ -41,7 +43,8 @@ class MSMainViewController: UITableViewController, AppCenterProtocol {
   private var startupModePicker: MSEnumPicker<StartupMode>?
   private var eventFilterStarted = false
   private var dbFileDescriptor: CInt = 0
-  private var dbFileSource: DispatchSourceProtocol?  
+  private var dbFileSource: DispatchSourceProtocol?
+
   let startUpModeForCurrentSession: NSInteger = (UserDefaults.standard.object(forKey: kMSStartTargetKey) ?? 0) as! NSInteger
   
   deinit {
@@ -72,6 +75,12 @@ class MSMainViewController: UITableViewController, AppCenterProtocol {
     // Otherwise channelGroup would be nil.
     if (StartupMode.allValues[startupMode] != .Skip && StartupMode.allValues[startupMode] != .None) {
         _ = MSTransmissionTargets.shared
+    }
+    
+    if let msaUserId = UserDefaults.standard.string(forKey: kMSATokenKey),
+        let refreshToken = UserDefaults.standard.string(forKey: kMSARefreshTokenKey) {
+        let provider = MSAnalyticsAuthenticationProvider(authenticationType: .msaCompact, ticketKey: msaUserId, delegate: MSAAnalyticsAuthenticationProvider.getInstance(refreshToken, self))
+        MSAnalyticsTransmissionTarget.addAuthenticationProvider(authenticationProvider:provider)
     }
 
     // Storage size section.